pyproject.toml
[project]
name = "curso-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.10",
    "fastapi>=0.124.4",
    "groq>=0.37.1",
    "ipykernel>=7.1.0",
    "openai>=2.12.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.14",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
